+++ extensions/disk-util/CHANGELOG.md
@@ -2,6 +2,5 @@
 
 ## [Initial Version] - {PR_MERGE_DATE}
 
-- Adds "Quick Create Volume" command to create APFS encrypted and case-sensitive volumes.
-- Adds "Create Volume" command to create APFS encrypted and case-sensitive volumes with extra features like symlinks (more features can be added later).
+- Adds "Create Volume" command to create APFS encrypted and case-sensitive volumes with extra features like symlinks etc.
 - Adds "Show Volumes" command to get details for volumes and perform operations.